L6 Move cardsMassive in separate js
